Revert drop on move, fixes drag cursor, improves sizing accuracy + refactor
